Task 6

+++ 0x00-hello_world/6-size.c
@@ -5,11 +5,11 @@
  */
 int main(void)
 {
-	printf("size of a char: %lu byte(S)\n", (unsigned long)sizeof(char));
-	printf("Size of an int: %lu byte(S)\n", (unsigned long)sizeof(int));
-	printf("Size of a long int: %lu byte(S)\n", (unsigned long)sizeof(long int));
-	printf("Size of a long long int: %lu byte(S)\n", (unsigned long)sizeof(long long int));
-	printf("Size of a float: %lu byte(S)\n", (unsigned long)sizeof(float));
+	printf("size of a char: %lu byte(S)\n", sizeof(char));
+	printf("Size of an int: %lu byte(S)\n", sizeof(int));
+	printf("Size of a long int: %lu byte(S)\n", sizeof(long int));
+	printf("Size of a long long int: %lu byte(S)\n", sizeof(long long int));
+	printf("Size of a float: %lu byte(S)\n", sizeof(float));
 	return (0);
 }
